+++ src/components/its/ManualImportTab.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent, RefObject, useState } from 'react';
 import { Box, Button, CircularProgress, Alert, SxProps, Theme } from '@mui/material';
 import { CommandResponse, Colors } from './ImportClusters';
-import OnboardingLogsDisplay from './OnboardingLogsDisplay';
-import CancelButton from './common/CancelButton';
+import OnboardingLogsDisplay from '../OnboardingLogsDisplay';
+import CancelButton from '../common/CancelButton';
 
 interface ManualImportTabProps {
   theme: string;
